api-gateway/requirements.txt
python==3.11.*
fastapi==0.115.*
uvicorn[standard]==0.34.*
pydantic==2.10.*
pydantic-settings==2.7.*
httpx==0.28.*
structlog==24.4.*
prometheus-client==0.21.*
redis==5.2.*
asyncpg==0.30.*
sqlalchemy[asyncio]==2.0.*
pymilvus==2.4.*
minio==7.2.*
langfuse==2.56.*
python-multipart==0.0.18
python-jose[cryptography]==3.3.*
passlib[bcrypt]==1.7.*
langdetect==1.0.9
pytest==8.0.*
pytest-asyncio==0.24.*
httpx-mock==0.32.*
